Refactor component folder structure for jobs and queues
