+++ client/src/components/dashboard/quick-actions.tsx
@@ -1,6 +1,6 @@
 import { Plus, RefreshCw, Trash2, Zap } from 'lucide-react';
 import React from 'react';
-import Card from '../shared/Card';
+import Card from '../shared/card-ui';
 
 interface QuickActionsProps {
     onRefresh: () => void;
