+++ client/src/components/dashboard/stat-card.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Card from '../shared/Card';
+import Card from '../shared/card-ui';
 
 interface StatCardProps {
     title: string;
