+++ client/src/components/queues/queues-summary.tsx
@@ -1,6 +1,6 @@
 import { BarChart3, Tag } from 'lucide-react';
 import React from 'react';
-import Card from '../shared/Card';
+import Card from '../shared/card-ui';
 
 interface QueueSummaryItem {
     name: string;
